+++ fiddles/nativeScript/fiddle-0000-Template/tools/tasks/seed/serve.docs.ts
@@ -1,0 +1,8 @@
+// import { serveDocs } from '../../utils';
+//
+// /**
+//  * Executes the build process, serving the application documentation using an
+//  * `express` server.
+//  */
+// export = serveDocs;
+