+++ scripts/bin/_utils.sh
@@ -159,4 +159,15 @@ function voidGrantAdmin() {
     sudo chown -R $(whoami):admin ${target} || exit $?;
 }
 
+function contains() {
+    string="$1"
+    substring="$2"
+    if test "${string#*$substring}" != "$string"
+    then
+        return 0    # $substring is in $string
+    else
+        return 1    # $substring is not in $string
+    fi
+}
+